# backend/cmd/app/docs/swagger.yaml
basePath: /
definitions:
...
  normaladmin_backend_internal_models.Admin:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      last_login_time:
        description: 改为指针类型，允许为 null
        type: string
      phone:
        type: string
      real_name:
        type: string
      role:
        $ref: '#/definitions/normaladmin_backend_internal_models.Role'
      role_id:
        type: integer
      status:
        description: 1-启用 0-禁用
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  normaladmin_backend_internal_models.Role:
    properties:
      code:
        type: string
      data_scope:
        description: 数据权限范围
        type: integer
      description:
        type: string
      id:
        description: 将 ID 字段的 JSON 标签设置为 "id"
        type: integer
      is_preset:
        description: 是否为系统预设角色
        type: boolean
      name:
        type: string
      remark:
        description: 备注
        type: string
      sort:
        description: 排序
        type: integer
      status:
        description: 状态
        type: integer
    type: object
  normaladmin_backend_pkg_utils_response.ResponseData:
    properties:
      data: {}
      error:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a backend server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Backend API
  version: "1.0"
paths:
...
  /gam/admins:
    get:
      consumes:
      - application/json
      description: 获取管理员列表，支持按用户名搜索、分页和排序
      parameters:
      - description: 用户名，用于搜索过滤
        in: query
        name: username
        type: string
      - default: 1
        description: 页码，默认值：1
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量，默认值：10
        in: query
        name: pageSize
        type: integer
      - description: 排序字段，多个字段用逗号分隔
        example: '"id,created_at"'
        in: query
        name: sortField
        type: string
      - description: 排序方式，多个排序方式用逗号分隔(asc/desc)
        example: '"desc,asc"'
        in: query
        name: sortOrder
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
            - properties:
                data:
                  properties:
                    admins:
                      items:
                        $ref: '#/definitions/normaladmin_backend_internal_models.Admin'
                      type: array
                    total:
                      type: integer
                  type: object
              type: object
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
      summary: 获取管理员列表
      tags:
      - 管理员管理
    post:
      consumes:
      - application/json
      description: 创建新的管理员账号，包括用户名、密码、角色等信息
      parameters:
      - description: 管理员信息，包括用户名、密码、角色ID等
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/normaladmin_backend_internal_models.Admin'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
            - properties:
                data:
                  properties:
                    admin:
                      $ref: '#/definitions/normaladmin_backend_internal_models.Admin'
                  type: object
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
      summary: 创建管理员
      tags:
      - 管理员管理
  /gam/admins/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除指定管理员账号
      parameters:
      - description: 管理员ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
            - properties:
                data:
                  properties:
                    message:
                      type: string
                  type: object
              type: object
        "400":
          description: 无效的管理员ID
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "404":
          description: 管理员不存在
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
      summary: 删除管理员
      tags:
      - 管理员管理
    get:
      consumes:
      - application/json
      description: 根据管理员ID获取详细信息
      parameters:
      - description: 管理员ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
            - properties:
                data:
                  properties:
                    admin:
                      $ref: '#/definitions/normaladmin_backend_internal_models.Admin'
                  type: object
              type: object
        "400":
          description: 无效的管理员ID
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "404":
          description: 管理员不存在
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
      summary: 获取单个管理员
      tags:
      - 管理员管理
    put:
      consumes:
      - application/json
      description: 更新管理员信息，包括用户名、角色等（密码单独更新）
      parameters:
      - description: 管理员ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: 管理员信息
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/normaladmin_backend_internal_models.Admin'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
            - properties:
                data:
                  properties:
                    admin:
                      $ref: '#/definitions/normaladmin_backend_internal_models.Admin'
                  type: object
              type: object
        "400":
          description: 无效的请求参数
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "404":
          description: 管理员不存在
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
      summary: 更新管理员
      tags:
      - 管理员管理
  /gam/admins/{id}/password:
    put:
      consumes:
      - application/json
      description: 更新管理员的登录密码，需要提供旧密码
      parameters:
      - description: 管理员ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: 密码信息
        in: body
        name: password
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
            - properties:
                data:
                  properties:
                    message:
                      type: string
                  type: object
              type: object
        "400":
          description: 无效的请求参数
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "401":
          description: 旧密码错误
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "404":
          description: 管理员不存在
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
      summary: 更新管理员密码
      tags:
      - 管理员管理
  /gam/admins/{id}/status:
    put:
      consumes:
      - application/json
      description: 更新管理员的启用/禁用状态（0：禁用，1：启用）
      parameters:
      - description: 管理员ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: 状态信息
        in: body
        name: status
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
            - properties:
                data:
                  properties:
                    message:
                      type: string
                  type: object
              type: object
        "400":
          description: 无效的请求参数
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "404":
          description: 管理员不存在
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
      summary: 更新管理员状态
      tags:
      - 管理员管理
  /gam/admins/check-field:
    get:
      consumes:
      - application/json
      description: 检查指定字段的值在管理员表中是否唯一
      parameters:
      - description: 字段名称
        example: '"username"'
        in: query
        name: field
        required: true
        type: string
      - description: 字段值
        example: '"admin"'
        in: query
        name: value
        required: true
        type: string
      - description: 需要排除的管理员ID
        in: query
        minimum: 1
        name: excludeId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
            - properties:
                data:
                  properties:
                    unique:
                      type: boolean
                  type: object
              type: object
        "400":
          description: 无效的请求参数
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
      summary: 检查管理员字段唯一性
      tags:
      - 管理员管理
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"